Fix OMZ permission warning on WSL
When dotfiles are stored outside of the WSL filesystem (lxfs), e.g.
for easy access by editors, oh-my-zsh will complain about permissions
on the completion scripts and refuse to parse them until you fix the
perms.  However files on windows fs mounted into WSL have their own
permission issues, e.g. everything is 755, so we make OMZ ignore that
in the case where we're both on WSL and the dotfiles are stored on a
non-lxfs filesustem (this may be brittle in the fiture?)